+++ .env
@@ -2,4 +2,5 @@ OPPLAERING=true
 MOCK_BACKEND=true
 ENVIRONMENT=dev
 DECORATOR_ENV=dev
-
+DINE_SYKMELDTE_ROOT=https://www-gcp.dev.nav.no/sykefravaerarbeidsgiver
+DITT_SYKEFRAVAER_ROOT=https://www-gcp.dev.nav.no/syk/sykefravaer
