+++ src/common/breadcrumbs/BreadcrumbsAppenderSM.tsx
@@ -1,0 +1,14 @@
+import { ReactElement, useEffect } from "react";
+import { useRouter } from "next/router";
+import { setBreadcrumbs } from "@navikt/nav-dekoratoren-moduler";
+import { createBreadcrumbsSM } from "@/common/breadcrumbs/index";
+
+export const BreadcrumbsAppenderSM = (): ReactElement => {
+  const { pathname } = useRouter();
+
+  useEffect(() => {
+    setBreadcrumbs(createBreadcrumbsSM(pathname));
+  }, [pathname]);
+
+  return <></>;
+};
