+++ src/pages/_app.tsx
@@ -1,11 +1,14 @@
 import "../styles/globals.css";
-import type { AppProps, AppContext } from "next/app";
+import type { AppContext, AppProps } from "next/app";
+import App from "next/app";
 import { QueryClient, QueryClientProvider } from "react-query";
 import { ReactQueryDevtools } from "react-query/devtools";
 import styled, { createGlobalStyle } from "styled-components";
 import { initAmplitude } from "@/common/amplitude/amplitude";
 import { useEffect } from "react";
-import App from "next/app";
+import { useAudience } from "@/common/hooks/routeHooks";
+import { BreadcrumbsAppenderSM } from "@/common/breadcrumbs/BreadcrumbsAppenderSM";
+import { BreadcrumbsAppenderAG } from "@/common/breadcrumbs/BreadcrumbsAppenderAG";
 
 const GlobalStyle = createGlobalStyle`
   body {
@@ -34,6 +37,8 @@ const minutesToMillis = (minutes: number) => {
 };
 
 function MyApp({ Component, pageProps }: AppProps) {
+  const { isAudienceSykmeldt } = useAudience();
+
   useEffect(() => {
     initAmplitude();
   }, []);
@@ -51,6 +56,11 @@ function MyApp({ Component, pageProps }: AppProps) {
   return (
     <QueryClientProvider client={queryClient}>
       <GlobalStyle />
+      {isAudienceSykmeldt ? (
+        <BreadcrumbsAppenderSM />
+      ) : (
+        <BreadcrumbsAppenderAG />
+      )}
       <ContentWrapperStyled>
         <Component {...pageProps} />
       </ContentWrapperStyled>
