+++ src/pages/arbeidsgiver/[narmestelederid]/index.tsx
@@ -10,7 +10,6 @@ import type { NextPage } from "next";
 import Head from "next/head";
 import { useDialogmoteDataAG } from "@/common/api/queries/arbeidsgiver/dialogmoteDataQueryAG";
 import MoteinnkallingPanel from "@/common/components/moteinnkalling/MoteinnkallingPanel";
-import { useBreadcrumbsAG } from "@/common/hooks/useBreadcrumbs";
 
 const texts = {
   title: "Dialogmøter",
@@ -19,7 +18,6 @@ const texts = {
 const Home: NextPage = () => {
   const narmestelederid = useNarmesteLederId();
   const dialogmoteData = useDialogmoteDataAG(narmestelederid);
-  useBreadcrumbsAG();
 
   if (dialogmoteData.isError) {
     return <div>Her ble det noe feil</div>;
